change project file structure
