+++ .gitignore
@@ -2,3 +2,4 @@
 __pycache__
 migrations
 config.py
+scripttorun.txt
