+++ app/app.py
@@ -1,0 +1,17 @@
+from main import create_app,db
+from flask_migrate import Migrate
+from main.controller.admin_route import admin
+from main.controller.gsm_cell import cell
+from main.controller.wifi_network import wifi
+
+app = create_app()
+
+migrate = Migrate(app, db)
+
+app.register_blueprint(admin,url_prefix="/admin")
+app.register_blueprint(cell,url_prefix="/celltower")
+app.register_blueprint(wifi,url_prefix="/wifi")
+
+if __name__ == "__main__":
+    app.run(host="0.0.0.0",port=5000)
+
