+++ app/main/__init__.py
@@ -1,0 +1,14 @@
+#from os import getenv
+#from dotenv import load_dotenv
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+
+db = SQLAlchemy()
+
+
+def create_app():
+    app = Flask(__name__)
+    app.config.from_object('config.Config')
+    db.init_app(app)
+    return app
